+++ mslicer/src/windows/sliced.rs
@@ -38,7 +38,7 @@ use crate::{
         },
     },
     render::slice_preview::SlicePreviewRenderCallback,
-    task::{FileDialog, IslandDetection, ReconstructMesh, SaveResult, TaskManager},
+    task::{FileDialog, IslandDetection, ReconstructMesh, SaveSliced, TaskManager},
     ui::{
         components::{collapsing_toggle, grid},
         management::{LazyText, LazyTextureId},
@@ -64,7 +64,7 @@ const DETECT_ISLANDS_DESC: &str =
     "Will color disconnected chunks of voxels red in the slice preview.";
 const SURFACE_AREA_DESC: &str = "Surface area in cm² of each layer. Layers with higher areas will adhere more to the FEP potentially causing print failures.";
 
-pub fn ui(app: &mut App, ui: &mut Ui, ctx: &Context) {
+pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
     if let Some(slice_operation) = &app.slice_operation {
         let progress = &slice_operation.progress;
 
@@ -218,7 +218,6 @@ pub fn ui(app: &mut App, ui: &mut Ui, ctx: &Context) {
                             &mut app.tasks,
                             &mut app.popup,
                             ui,
-                            ctx,
                         );
                     })
                 });
@@ -504,7 +503,7 @@ fn save_file(
             let file_name = path.file_name().unwrap().to_string_lossy();
             let mut out = File::create(&path).unwrap();
 
-            tasks.push(Box::new(SaveResult::new(
+            tasks.push(Box::new(SaveSliced::new(
                 (format, data.clone(), config, preview),
                 file_name.into_owned(),
                 move |bytes| out.write_all(&bytes).unwrap(),
@@ -546,7 +545,6 @@ fn sidebar(
     tasks: &mut TaskManager,
     popups: &mut PopupManager,
     ui: &mut Ui,
-    ctx: &Context,
 ) {
     CollapsingHeader::new("Preview Image")
         .default_open(true)
@@ -634,7 +632,7 @@ fn sidebar(
             let (width, height) = (preview.image.width(), preview.image.height());
 
             let size = vec2(available, available / width as f32 * height as f32);
-            let texture = SizedTexture::new(preview.texture.get(ctx, &preview.image), size);
+            let texture = SizedTexture::new(preview.texture.get(ui.ctx(), &preview.image), size);
 
             reset_preview.then(|| previews.take());